# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  telegram.Alert:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      endsAt:
        type: string
      fingerprint:
        type: string
      generatorURL:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      startsAt:
        type: string
      status:
        type: string
    type: object
  telegram.AlertManagerWebhook:
    properties:
      alerts:
        items:
          $ref: '#/definitions/telegram.Alert'
        type: array
      commonAnnotations:
        additionalProperties:
          type: string
        type: object
      commonLabels:
        additionalProperties:
          type: string
        type: object
      externalURL:
        type: string
      groupKey:
        type: string
      groupLabels:
        additionalProperties:
          type: string
        type: object
      receiver:
        type: string
      status:
        type: string
      truncatedAlerts:
        type: integer
      version:
        type: string
    type: object
  telegram.SendMessageRequest:
    properties:
      alertmanager_data:
        allOf:
        - $ref: '#/definitions/telegram.AlertManagerWebhook'
        description: AlertManager webhook 數據
      message:
        description: 可選，當使用 AlertManager 格式時不需要
        type: string
      template_language:
        description: 可選，優先使用配置檔案中的設定
        type: string
    type: object
  telegram.SendMessageResponse:
    properties:
      level:
        type: integer
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    email: vincent119@gmail.com
    name: API Support
  description: 這是 Alert Webhooks 的 Swagger 文件（由 swag 自動產生）。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Alert Webhooks API
  version: "1.0"
paths:
...
  /telegram/chatid_{chatid}:
    post:
      consumes:
      - application/json
      description: Send message to specified Telegram chat level
      parameters:
      - description: '聊天等級 (格式: L{0-4})'
        in: path
        name: chatid
        required: true
        type: string
      - description: 訊息內容
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/telegram.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/telegram.SendMessageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: Send Telegram message
      tags:
      - telegram
  /telegram/info:
    get:
      description: 獲取 Telegram 機器人的基本資訊
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BasicAuth: []
      summary: 獲取機器人資訊
      tags:
      - telegram
produces:
- application/json
schemes:
- https
- http
security:
- BasicAuth: []
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"